separate drawer and contents
