+++ client/src/main/scripts/components/homepage/DrawerContents.tsx
@@ -17,30 +17,26 @@
  *
  * You can contact us at team@mytake.org
  */
-import * as React from "react";
-import * as renderer from "react-test-renderer";
-import Drawer from "./Drawer";
-import { INFO_HEADER_TABS_ENUM } from "./infoHeader";
+import React from "react";
+import GetInvolved from "./GetInvolved";
+import HowToUseThis from "./HowToUseThis";
+import { INFO_HEADER_TABS_ENUM, useSocials } from "./infoHeader";
+import WhatIsThis from "./WhatIsThis";
 
-jest.mock("./DrawerContents", () => ({
-  __esModule: true,
-  default: "DrawerContents",
-}));
+interface DrawerContentsProps {
+  activeTab: INFO_HEADER_TABS_ENUM;
+}
 
-test("Drawer - expanded", () => {
-  const props = {
-    activeTab: INFO_HEADER_TABS_ENUM.HOW_TO_USE_THIS,
-    isExpanded: true,
-  };
-  const tree = renderer.create(<Drawer {...props} />).toJSON();
-  expect(tree).toMatchSnapshot();
-});
+const DrawerContents: React.FC<DrawerContentsProps> = (props) => {
+  const { leftSocial, rightSocial } = useSocials();
+  switch (props.activeTab) {
+    case INFO_HEADER_TABS_ENUM.GET_INVOLVED:
+      return <GetInvolved />;
+    case INFO_HEADER_TABS_ENUM.HOW_TO_USE_THIS:
+      return <HowToUseThis />;
+    case INFO_HEADER_TABS_ENUM.WHAT_IS_THIS:
+      return <WhatIsThis leftSocial={leftSocial} rightSocial={rightSocial} />;
+  }
+};
 
-test("Drawer - collapsed", () => {
-  const props = {
-    activeTab: INFO_HEADER_TABS_ENUM.HOW_TO_USE_THIS,
-    isExpanded: false,
-  };
-  const tree = renderer.create(<Drawer {...props} />).toJSON();
-  expect(tree).toMatchSnapshot();
-});
+export default DrawerContents;
